refactor(metric-reader): move callWithTimeout and TimeoutError to src/utils.ts
