+++ experimental/packages/opentelemetry-sdk-metrics-base/src/utils.ts
@@ -38,3 +38,44 @@ export function hashAttributes(attributes: Attributes): string {
     return (result += key + ':' + attributes[key]);
   }, '|#');
 }
+
+/**
+ * Error that is thrown on timeouts.
+ */
+export class TimeoutError extends Error {
+  constructor(message?: string) {
+    super(message);
+    Object.setPrototypeOf(this, TimeoutError.prototype);
+  }
+}
+
+/**
+ * Adds a timeout to a promise and rejects if the specified timeout has elapsed. Also rejects if the specified promise
+ * rejects, and resolves if the specified promise resolves.
+ *
+ * <p> NOTE: this operation will continue even after it throws a {@link TimeoutError}.
+ *
+ * @param promise promise to use with timeout.
+ * @param timeout the timeout in milliseconds until the returned promise is rejected.
+ */
+export function callWithTimeout<T>(promise: Promise<T>, timeout: number): Promise<T> {
+  let timeoutHandle: ReturnType<typeof setTimeout>;
+
+  const timeoutPromise = new Promise<never>(function timeoutFunction(_resolve, reject) {
+    timeoutHandle = setTimeout(
+      function timeoutHandler() {
+        reject(new TimeoutError('Operation timed out.'));
+      },
+      timeout
+    );
+  });
+
+  return Promise.race([promise, timeoutPromise]).then(result => {
+      clearTimeout(timeoutHandle);
+      return result;
+    },
+    reason => {
+      clearTimeout(timeoutHandle);
+      throw reason;
+    });
+}
